Changed MessageInfo.

+++ mona/include/sys/types.h
@@ -154,11 +154,11 @@ typedef struct {
 #define MESSAGE_INFO_MAX_STR_LENGTH 128
 
 typedef struct {
-    uint32_t header;
-    uint32_t arg1;
-    uint32_t arg2;
-    uint32_t arg3;
-    uint32_t from;
+    uintptr_t header;
+    uintptr_t arg1;
+    uintptr_t arg2;
+    uintptr_t arg3;
+    uintptr_t from;
     char str[MESSAGE_INFO_MAX_STR_LENGTH];
     int length;
 } MessageInfo;
